Fixed tests that were broken by work on SAMC
git-svn-id: https://phycas.svn.sourceforge.net/svnroot/phycas/trunk@492 b774380d-c90d-0410-997c-aec342e5d6fd

+++ phycas/src/basic_tree.inl
@@ -299,7 +299,7 @@ inline unsigned Tree::GetNNodes()
 inline unsigned Tree::GetNInternalsAllocated()
 	{
 	RefreshNodeCounts(); //temp
-	return internalNodeStorage.size() + GetNInternals();
+	return (unsigned)(internalNodeStorage.size() + GetNInternals());
 	}
